requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
branca==0.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudconvert==2.1.0
exceptiongroup==1.2.0
folium==0.16.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
polyline==2.0.2
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
xyzservices==2023.10.1
yarl==1.9.4
